Removed the small style puzzle idea puzzles will instead be on a macro scale

+++ in-the-city/src/main/java/com/nicktagliamonte/items/Item.java
@@ -12,7 +12,7 @@ public class Item {
     @Expose private boolean isConsumable;
     @Expose private double value;
 
-    public Item(String name, String description, double weight, boolean isConsumable, int value) {
+    public Item(String name, String description, double weight, boolean isConsumable, double value) {
         this.name = name;
         this.description = description;
         this.weight = weight;
